+++ index.html
@@ -1,24 +1,19 @@
 <!DOCTYPE html>
 <html>
+<head>
+  <meta charset="utf-8">
+  <title>Breakout Bastards</title>
+</head>
+<body>
+  <canvas id="game"
+          class="game-canvas"
+          height="500"
+          width="700"
+          style="border: 5px solid black">
+  </canvas>
 
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=yes">
-    <link rel="stylesheet" type="text/css" href="css/reset.css">
-      <link rel="stylesheet" type="text/css" href="css/styles.css">
-    <title>Breakout Bastards</title>
-  </head>
-
-  <body>
-    <header class="header">Breakout Bastards</header>
-    <canvas id="game"
-            class="game-canvas"
-            height="500"
-            width="700">
-    </canvas>
-  </body>
-
-  <script src="lib/gametime.js"></script>
+  <script src="lib/index.js"></script>
   <script src="main.bundle.js"></script>
   <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
+</body>
 </html>
