+++ lib/brick.js
@@ -1,0 +1,6 @@
+function Brick (x, y) {
+  this.x = x;
+  this.y = y;
+}
+
+module.exports = Brick;
